+++ project/src/main/resources/templates/board/detail.html
@@ -7,6 +7,7 @@
     <h1 th:text="${board.title}"></h1>
     <p th:text="${board.content}"></p>
     <!-- <small>작성일: <span th:text="${#dates.format(board.createdAt, 'yyyy-MM-dd HH:mm')}"></span></small> -->
+    <small>작성일: <span th:text="${#temporals.format(board.createdAt, 'yyyy-MM-dd HH:mm')}"></span></small>
     <br />
     <a th:href="@{/board/{id}/edit(id=${board.id})}">수정</a>
     <form th:action="@{/board/{id}/delete(id=${board.id})}" method="post" style="display:inline;">
